LayoutTests/imported/w3c/web-platform-tests/wai-aria/role/listbox-roles.tentative-expected.txt

x

FAIL orphaned option outside the context of listbox assert_equals: <nav role="option" data-testname="orphaned option outside the context of listbox" data-expectedrole="navigation" class="ex">x
</nav> expected "navigation" but got "option"
